add vote authentication

+++ app/controllers/index.rb
@@ -10,10 +10,14 @@ get '/questions/:id' do
 end
 
 post '/questions/:question_id/answers/:answer_id/upvote' do
-  answer = Answer.find(params[:answer_id])
-  answer.scores.create(value: 1)
-  score = answer.scores.count
-  score.to_s
+  if session[:user_id]
+    answer = Answer.find(params[:answer_id])
+    answer.scores.create(value: 1)
+    score = answer.scores.count
+    score.to_s
+  else
+    return 'redirect'
+  end
 end
 
 get '/question/new' do
@@ -27,8 +31,12 @@ post '/questions' do
 end
 
 post '/questions/:question_id/answers/:answer_id/downvote' do
-  answer = Answer.find(params[:answer_id])
-  answer.scores.last.destroy
-  score = answer.scores.count
-  score.to_s
+  if session[:user_id]
+    answer = Answer.find(params[:answer_id])
+    answer.scores.last.destroy
+    score = answer.scores.count
+    score.to_s
+  else
+    return 'redirect'
+  end
 end
